Added Pew Pew with example scene
Added some pretty rushed code for gun shooty didn't have time for UI sorry, if you want to do ui and need the ammo info change the ammoInMag and ammoInReserve in GunScript.cs to public. Aiming is still a little weird if you want to fix it be my  guest.  The test scene included shows how all the stuff should work that's basically about it.